product/serializers.py
from rest_framework import serializers

from .models import Catagory, Product
class CategorySerializer(serializers.ModelSerializer):
    class Meta:
        model = Catagory
        fields = '__all__'

class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = ['id', 'name', 'catagory', 'price', 'quantity']